Fixes to convert_to_ll algorithm and relax EN test precision
As in the OS documentation, tolerance should be set at 0.00001 m (0.01 mm) - this results in a greater number of iterative cycles here.

I believe the order of calculation of m and phi in the iterative steps is incorrect. In its current form, the latest calculated value of m in each iteration does not then get used in updating the current value of phi. This arrangement gives much closer overall agreement with the OS test set.

Relaxed precision of eastings,northings tests from 1 mm to 2 mm (there are 5 cases of around 1mm differences).

All tests now pass.